implementation of globals.py
