+++ lib/doconce/rst.py
@@ -8,7 +8,8 @@ from .common import insert_code_and_tex, indent_lines, \
     table_analysis, plain_exercise, bibliography, \
     cite_with_multiple_args2multiple_cites, fix_ref_section_chapter
 from .html import html_movie, html_quiz
-from .doconce import _abort, errwarn, locale_dict
+from .doconce import _abort, errwarn
+from doconce import globals
 from .misc import option, _abort
 
 def rst_abstract(m):
@@ -723,7 +724,7 @@ def define(FILENAME_EXTENSION,
 
     TABLE['rst'] = rst_table
     EXERCISE['rst'] = plain_exercise
-    TOC['rst'] = lambda s, f: '.. contents:: %s\n   :depth: 2' % locale_dict[locale_dict['language']].get('toc', 'Table of contents')
+    TOC['rst'] = lambda s, f: '.. contents:: %s\n   :depth: 2' % globals.locale_dict[globals.locale_dict['language']].get('toc', 'Table of contents')
     QUIZ['rst'] = rst_quiz
     INTRO['rst'] = """\
 .. Automatically generated reStructuredText file from DocOnce source
